requirements.txt
torch>=2.1.0
torchvision>=0.15.0
transformers>=4.40.0
Pillow>=9.5.0
pillow-heif>=0.13.0
numpy<2
